change to float

+++ core/src/com/mygdx/game/entityComponentSystem/components/Size.java
@@ -3,6 +3,6 @@ package com.mygdx.game.entityComponentSystem.components;
 import com.badlogic.ashley.core.Component;
 
 public class Size implements Component {
-    public int width = 32;
-    public int height = 32;
+    public float width = 32;
+    public float height = 32;
 }
